[lookon-sys] 커뮤니티 및 피드 detail : 상세병견내용
