+++ ks55team02/src/main/java/ks55team02/customer/feed/controller/FeedController.java
@@ -1,16 +1,12 @@
 package ks55team02.customer.feed.controller;
 
-import java.util.Map;
-
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.SessionAttribute;
 
-import jakarta.servlet.http.HttpSession;
 import ks55team02.customer.feed.domain.Feed;
 import ks55team02.customer.feed.service.FeedService;
 import ks55team02.customer.login.domain.LoginUser;
@@ -50,16 +46,16 @@ public class FeedController {
 	     return "customer/feed/feedDetail";
 	 }
 	    
-    // 마이 피드 조회
-    @GetMapping("/feedListByMe")
-    public String selectFeedListByMe(
+    // 나의 마이 피드 조회
+    @GetMapping("/feedListByUserNo")
+    public String selectFeedListByUserNo(
             @SessionAttribute(name = "loginUser", required = false) LoginUser loginUser, 
             Model model) {
     	
     	if(loginUser == null) {
             // @SessionAttribute(required = false)를 사용하면 세션이 없을 때 null이 주입됩니다.
             // 기존과 같이 로그인 페이지로 리다이렉트합니다.
-    		return "redirect:/customer/login/login?redirectUrl=/customer/feed/feedListByMe";
+    		return "redirect:/customer/login/login?redirectUrl=/customer/feed/feedListByUserNo";
     	}
     	
     	String userNo = loginUser.getUserNo();
@@ -69,8 +65,17 @@ public class FeedController {
     	
         // feedList와 hasNext는 더 이상 모델에 추가하지 않습니다.
     	
-    	return "customer/feed/feedListByMe";
+    	return "customer/feed/feedListByUserNo";
     }
     
+    // 타인의 마이 피드 조회
+    @GetMapping("/feedListByUserNo/{userNo}")
+    public String userFeedPage(@PathVariable("userNo") String userNo, Model model) {
+        // 경로 변수(userNo)를 이용해 타인의 프로필 정보를 조회합니다.
+        UserInfoResponse userInfo = userInfoService.getUserInfo(userNo);
+        model.addAttribute("userInfo", userInfo);
+        // "마이피드"와 "타인피드"는 동일한 뷰 파일을 재사용합니다.
+        return "customer/feed/feedListByUserNo";
+    }
 
 }
