+++ ks55team02/src/main/resources/templates/customer/feed/feedDetail.html
@@ -24,8 +24,8 @@
             <!-- 1. 작성자 정보 및 팔로우/더보기 버튼 -->
             <header class="feed-post-header">
                 <div class="user-profile">
-                    <a href="#">
-                        <img th:src="*{writerInfo?.prflImg != null ? writerInfo.prflImg : '/images/default-profile.png'}" alt="프로필 사진" class="profile-image">
+                    <a th:href="@{/customer/feed/feedListByUserNo/{userNo}(userNo=*{writerInfo.userNo})}">
+                        <img th:src="*{writerInfo?.prflImg != null ? writerInfo.prflImg : '/uploads/profiles/profiles.jpg'}" alt="프로필 사진" class="profile-image">
                     </a>
                     <div class="user-info">
                         <span class="username" th:text="*{writerInfo?.userNcnm}">유저닉네임</span>
