+++ ks55team02/src/main/resources/templates/customer/feed/feedList.html
@@ -37,7 +37,7 @@
 		                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M15.5 14h-.79l-.28-.27A6.471 6.471 0 0 0 16 9.5 6.5 6.5 0 1 0 9.5 16c1.61 0 3.09-.59 4.23-1.57l.27.28v.79l5 4.99L20.49 19l-4.99-5zm-6 0C7.01 14 5 11.99 5 9.5S7.01 5 9.5 5 14 7.01 14 9.5 11.99 14 9.5 14z"/></svg>
 		            </a>
 		            <!-- 마이피드 아이콘 버튼 -->
-		            <a th:href="@{/customer/feed/feedListByMe}"
+		            <a th:href="@{/customer/feed/feedListByUserNo}"
 		            	th:if="${session.loginUser != null}" class="action-btn" id="my-feed-btn" aria-label="마이피드">
 		                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4-4 1.79-4 4 1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z"/></svg>
 		            </a>
